feat(DDD with Clean Architeture) valida se foi o criados da questao, para poder deletar questao

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -3,12 +3,23 @@ import { Question } from '../../enterprise/entities/question'
 
 interface DeleteQuestionUseCaseRequest {
   question: Question
+  authorId: string
 }
 
 export class DeleteQuestionUseCase {
   constructor(private questionRepository: QuestionRepository) {}
 
-  async execute({ question }: DeleteQuestionUseCaseRequest) {
+  async execute({ question, authorId }: DeleteQuestionUseCaseRequest) {
+    await this.questionRepository.findById(question.id.toString()!)
+
+    if (!question) {
+      throw new Error('Question not found')
+    }
+
+    if (question.authorId.toString() !== authorId) {
+      throw new Error('You are not the author of this question')
+    }
+
     this.questionRepository.delete(question)
   }
 }
